+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve zh_name Field in EmployeeForm 
 
 ### Fixed
+- Missing `nullable=False` in Employee model
 - Typo in EmployeeForm
 
 ## [0.1.0] - 2021-01-04